+++ Changelog.md
@@ -5,6 +5,15 @@
 1. Move private headers from include/sdf to src folder.
     * [Pull request 553](https://bitbucket.org/osrf/sdformat/pull-requests/553)
 
+1. Fix ign library path on macOS.
+    * [Pull request 542](https://bitbucket.org/osrf/sdformat/pull-requests/542)
+
+1. Use `ign sdf --check` to check sibling elements of the same type for non-unique names.
+    * [Pull request 554](https://bitbucket.org/osrf/sdformat/pull-requests/554)
+
+1. Converter: remove all matching elements specified by `<remove>` tag.
+    * [Pull request 551](https://bitbucket.org/osrf/sdformat/pull-requests/551)
+
 ### SDFormat 8.3.0 (2019-08-17)
 
 1. Added Actor DOM
@@ -156,6 +165,9 @@
 
 ### SDFormat 7.0.0 (xxxx-xx-xx)
 
+1. Fix ign library path on macOS.
+    * [Pull request 542](https://bitbucket.org/osrf/sdformat/pull-requests/542)
+
 1. Preserve XML elements that are not part of the SDF specification.
     * [Pull request 449](https://bitbucket.org/osrf/sdformat/pull-requests/449)
 
@@ -201,6 +213,15 @@
 ## SDFormat 6.0
 
 ### SDFormat 6.X.X (20XX-XX-XX)
+
+1. Fix ign library path on macOS.
+    * [Pull request 552](https://bitbucket.org/osrf/sdformat/pull-requests/552)
+
+1. Use `ign sdf --check` to check sibling elements of the same type for non-unique names.
+    * [Pull request 554](https://bitbucket.org/osrf/sdformat/pull-requests/554)
+
+1. Converter: remove all matching elements specified by `<remove>` tag.
+    * [Pull request 551](https://bitbucket.org/osrf/sdformat/pull-requests/551)
 
 ### SDFormat 6.2.0 (2019-01-17)